This commit contains the files associated with Nathanael Mandy's thesis.
The folder named "Merge" contains a new version of the WMORE firmware built on v2.7 of the SparkFun firmware.

The "OLA Bluetooth" folder contains basic scripts to implement bluetooth as well as OLA firmware that can stream bluetooth values.

The "Laptop Bluetooth" folder contains python scripts that implement bluetooth on the laptop side to connect to the OLA.