Update WMORE_Openlog_Logger_2.10_draft.ino
In beginIMU
Removed input parameter "silent"
Set IMU SPI rate to 7MHz
Set IMU SPI rate to 7MHz
Removed if(!silent) condition for ICM-20948 failed to init
Replaced the if DMP code with a full startup for non-DMP mode, as if imuUseDMP == false

Undid the previous commit commenting that a print relating to SD card failed was not in OLAv2.10. It was, I just accidentally deleted the whole if(!silent) code block instead of just deleting the condition.